requirements.txt
steamlit
torch
Pillow
streamlit_drawable_canvas
